Better handle saving of files
First off, handle the case when the user hasn't logged in to Dropbox
yet. They will be sent to log in, and then they should be prompted for
the filename they want to use.

After a user chooses a filename, all saves following will use that same
name and create a new version in Dropbox. If a user clicks the clear
button, they will essentially be creating a new document, and they will
be prompted for a new filename when they click save.

Duplicate filenames are handled. If the user enters a name of a file
that is already in Dropbox, the Dropbox API will automatically create a
new file (probably with the requested filename + "(1)" or something
like that)